+++ game/cell.py
@@ -18,3 +18,4 @@ class Cell:
             return value
         else:
             return self.letter.value
+